+++ src/utils/array.js
@@ -29,6 +29,36 @@ export function moveFromTo<E>(
   return insertAt(newIndex, arr[oldIndex], without);
 }
 
+export function insertSpans<E>(
+  spans: $ReadOnlyArray<[number, $ReadOnlyArray<E>]>,
+  arr: $ReadOnlyArray<E>
+): Array<E> {
+  // no duplicated indices are allowed, ECMAScript Array.sort is not stable by spec
+  const indexSet = new Set(spans.map(([i]) => i));
+  if (indexSet.size !== spans.length) {
+    throw new Error(
+      "You cannot insert two spans at the same index. Combine the values of the spans."
+    );
+  }
+
+  // sort spans by insertion position
+  const spansCopy = [...spans];
+  spansCopy.sort(([i], [j]) => i - j);
+
+  // build the new array in one pass
+  let ret = [];
+  let lastIndexInsertedAt = 0;
+  spansCopy.forEach(([index, contents]) => {
+    // All the content before this
+    ret = ret.concat(arr.slice(lastIndexInsertedAt, index));
+    ret = ret.concat(contents);
+    lastIndexInsertedAt = index;
+  });
+  ret = ret.concat(arr.slice(lastIndexInsertedAt));
+
+  return ret;
+}
+
 // Strict on length
 export function zipWith<A, B, C>(
   f: (A, B) => C,
@@ -41,6 +71,25 @@ export function zipWith<A, B, C>(
   const ret = [];
   for (let i = 0; i < left.length; i += 1) {
     ret.push(f(left[i], right[i]));
+  }
+  return ret;
+}
+
+export function zip<A, B>(
+  left: $ReadOnlyArray<A>,
+  right: $ReadOnlyArray<B>
+): Array<[A, B]> {
+  return zipWith((l, r) => [l, r], left, right);
+}
+
+export function unzip<A, B>(
+  zipped: $ReadOnlyArray<[A, B]>
+): [Array<A>, Array<B>] {
+  const ret = [[], []];
+  for (let i = 0; i < zipped.length; i += 1) {
+    const [left, right] = zipped[i];
+    ret[0].push(left);
+    ret[1].push(right);
   }
   return ret;
 }